Add task manager to limit concurrency jobs for dump tool

+++ tools/dump/index.js
@@ -14,8 +14,89 @@ const regex = /(?:game(?:-\w+)?\.granbluefantasy\.jp)|(?:gbf\.game(?:-\w+)?\.mbg
 
 const targetFolder = process.env.targetFolder || 'collected'
 const applyPrettier = process.env.applyPrettier || false
+const versionFile = path.resolve(targetFolder, 'version.txt')
 
 const yarn = process.platform === 'win32' ? 'yarn.cmd' : 'yarn'
+const cpuCount = require('os').cpus().length
+const queue = {}
+const task = {}
+let assetVersion
+let fsIsReady = true
+
+try {
+  assetVersion = Number(fs.readFileSync(versionFile, 'utf8'))
+} catch (error) {}
+
+function taskRunner () {
+  if (fsIsReady && queue[assetVersion] && queue[assetVersion].length && Object.keys(task).length < cpuCount) {
+    // if queue is not empty and running task is lower than logic cpu count
+    const child = exec(
+      `${yarn} -s prettier-standard ${queue[assetVersion].shift()}`,
+      { windowsHide: true },
+      (error, stdout, stderr) => {
+        if (error) err(stderr)
+        if (stdout) log(stdout)
+      }
+    )
+
+    task[child.pid] = {
+      version: assetVersion,
+      kill: child.kill
+    }
+
+    child.once('exit', () => {
+      // Remove current task and re-run taskRunner after task finished
+      delete task[child.pid]
+      taskRunner()
+    })
+    // self fill all task slot
+    taskRunner()
+  }
+}
+
+function checkAssets (pathname) {
+  return new Promise(resolve => {
+    const extractFilename = /^assets\/(\d+)\/(.*)$/.exec(pathname)
+    if (extractFilename) {
+      const outputPath = path.resolve(targetFolder, 'currentAssets', extractFilename[2])
+      const currentVersion = Number(extractFilename[1])
+      if (assetVersion && assetVersion < currentVersion) {
+        // close all new incoming task
+        fsIsReady = false
+        // if new version pending, then clean all current assets, but not include static assets.
+        fs.emptyDirSync(path.resolve(targetFolder, 'currentAssets', 'js'))
+        fs.emptyDirSync(path.resolve(targetFolder, 'currentAssets', 'css'))
+        fs.emptyDirSync(path.resolve(targetFolder, 'prettified'))
+        // clean all pending queue and kill all current running task that belong to old version.
+        Object.keys(queue).forEach(version => {
+          if (version && version !== currentVersion) {
+            delete queue[version]
+          }
+        })
+        Object.keys(task).forEach(pid => {
+          if (pid && pid.version !== currentVersion) {
+            pid.kill()
+            delete task[pid]
+          }
+        })
+        // update version info
+        assetVersion = currentVersion
+        fs.writeFile(versionFile, currentVersion, error => { if (error) err(error) })
+        // file system is ready
+        fsIsReady = true
+      }
+      resolve({
+        extractFilename,
+        outputPath,
+        currentVersion
+      })
+    } else {
+      resolve({
+        outputPath: path.resolve(targetFolder, pathname.replace(/^assets\//, 'currentAssets/'))
+      })
+    }
+  })
+}
 
 // Inspect response with game content only
 proxy.setResInspectCondition((clientRequest, remoteResponse) => {
@@ -24,65 +105,38 @@ proxy.setResInspectCondition((clientRequest, remoteResponse) => {
 
 proxy.smartListen(Number(process.env.listenPort) || 8000)
 
-function promiseProcess (exists, name, ext, dir, response) {
-  let append = 0
-  return new Promise(resolve => {
-    if (exists) {
-      const filePattern = new RegExp(`${name}(?:-(\\d+))?${ext}`)
-      fs.readdirSync(dir).forEach(file => {
-        const result = filePattern.exec(path.basename(file))
-        if (result) {
-          if (fs.readFileSync(path.join(dir, file)).compare(response.body) !== 0) {
-            const num = Number(result[1] || 0)
-            num > append && (append = num + 1)
-          }
-        }
-      })
-      resolve(append)
-    } else {
-      resolve()
-    }
-  })
-}
-
 // get response body and save
 proxy.on('getResponse', (response) => {
   let { pathname } = urlParser(response.url)
   pathname === '/' && (pathname += 'index.html')
   pathname = pathname.replace(/^\//, '')
   if (pathname.indexOf('assets/') === 0 || pathname === 'index.html') {
-    const outputPath = path.resolve(targetFolder, pathname)
-    const { dir, name, ext } = path.parse(outputPath)
-    fs.pathExists(outputPath, (error, exists) => {
-      if (error) { err(error) }
-      promiseProcess(exists, name, ext, dir, response).then(append => {
-        const fileName = append ? `${name}-${append}${ext}` : `${name}${ext}`
-        fs.pathExists(path.resolve(dir, fileName), (error, exists) => {
-          if (error) { err(error) }
-          if (exists) {
-            log('  exists: %s', path.join(path.dirname(pathname), fileName))
-          } else {
-            log('new file: %s', path.join(path.dirname(pathname), fileName))
-            fs.outputFile(path.resolve(dir, fileName), response.body, (error) => {
+    checkAssets(pathname).then(result => {
+      fs.pathExists(result.outputPath, (error, exists) => {
+        if (error) { err(error) }
+        if (exists && fs.readFileSync(result.outputPath).compare(response.body) !== 0) {
+          log('Exists: %s', pathname)
+        } else {
+          log('Create: %s', pathname)
+          fs.outputFile(result.outputPath, response.body, (error) => {
+            if (error) { err(error) }
+          })
+          if (applyPrettier && result.extractFilename) {
+            const extractPath = path.resolve(targetFolder, 'prettified', result.extractFilename[2])
+            fs.outputFile(extractPath, response.body, (error) => {
               if (error) { err(error) }
+              if (queue[result.currentVersion] === undefined) {
+                queue[result.currentVersion] = []
+              }
+              queue[result.currentVersion].push(extractPath)
+              taskRunner()
             })
-            const extractFilename = /^assets\/\d+\/(.*)$/.exec(pathname)
-            if (applyPrettier && extractFilename) {
-              const extractPath = path.resolve(targetFolder, 'prettified', extractFilename[1])
-              fs.outputFile(extractPath, response.body, (error) => {
-                if (error) { err(error) }
-                exec(`${yarn} -s prettier-standard ${extractPath}`, { windowsHide: true }, (error, stdout, stderr) => {
-                  if (error) err(stderr)
-                  if (stdout) log(stdout)
-                })
-              })
-            }
           }
-        })
-      }).catch(reason => err(reason))
+        }
+      })
     })
   } else {
-    log('  bypass: %s', pathname)
+    log('Bypass: %s', pathname)
   }
 })
